Simplify complex Variant price behaviour
When I first implemented this, I thought that we would be moving to
explicit variants with differing prices quite soon.

If we ever move to variants with different prices to the master ( e.g.
variants with extra price for extra sizes or explicit variants for
custom colours ) then we will need to be cleverer about not assuming all
prices on all variants should be the same.