Curated: Allow only curated images to show.
After discussion with Toby and Tania, we need to allow an option where only the curated items are displayed.  Done this by adding a new 'variable' of "curated"=>true.

Note: if curated is set to true we force the total_products to equal the number of curated products to stop the "more dresses" button appearing.