Fix incorrect colour on curated images over offset.
Found a bug where curated images over the limit (the default is 21) used the colours from the first dresses in the list.  Example: 687-black number 22 in the PID list tried to display as 687-coral, the colour from the first PID, which is 471-coral.

Fixed by adding the offset to the colour index.