Add new task to reset PK sequences
When dealing with database dumps and dirty databases, the number of rows
in the table and the value of the primary key sequence can become
discombobulated.

If you get errors on duplicate ID on insert, simply run this rake task
to set the sequence to max(id) for all tables.