[fix] adjust mobile nav behaviour when clicking on items
This bug made the nav bar always animate later then the actual click. Example: loaded the page, clicked on the 4th nav item, the navbar would go to the initial position. Then clicked on the 6th item, it would go to the 4th item instead (because it was the last click).