[FEATURE][WED-30] Adds Moodboard chat
Added the moodboard chat componente created with React, and some styling for it.
Also some general buttons icons and styling for the moodboard which will be part of
the chat. Also added an endpoint to the rails app to get the twilio token to get
access to the api for the chat.